+++ tests/tests/test-01-mirtOne.R
@@ -32,7 +32,7 @@ test_that('dich', {
     expect_equal(extract.mirt(modm3, 'df'), 25)
     expect_equal(extract.mirt(modm3, 'condnum'), 4.741137, tolerance = 1e-4)
     LG <- lagrange(modm3, parnum = list(1, 5))
-    expect_equal(LG$X2, c(0.25555838, 0.02038625), tolerance = 1e-4)
+    expect_equal(LG$X2, c(0.46730208, 0.05686589), tolerance = 1e-4)
     dat <- expand.table(LSAT6)
     modm3 <- mirt(dat, 1, itemtype = 'Rasch', SE = TRUE, SE.type = 'SEM', verbose=FALSE)
     expect_is(modm3, 'SingleGroupClass')